feat: split select entry test
