--- tests/integration/server-config/tests/typings.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="@types/jest" />
