+++ tests/integration/server-config/tsconfig.json
@@ -5,24 +5,10 @@
     "jsx": "preserve",
     "baseUrl": "./",
     "paths": {
-      "@/*": [
-        "./src/*"
-      ],
-      "@shared/*": [
-        "./shared/*"
-      ],
-      "@api/*": [
-        "./api/*"
-      ]
+      "@/*": ["./src/*"],
+      "@shared/*": ["./shared/*"],
+      "@api/*": ["./api/*"]
     }
   },
-  "include": [
-    "src",
-    "shared",
-    "config",
-    "api",
-    "tests",
-    "plugins",
-    "server"
-  ]
+  "include": ["src", "shared", "config", "api", "plugins", "server"]
 }
