+++ tests/integration/server-hook/middleware/tsconfig.json
@@ -9,5 +9,5 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src", "test", "server", "modern.config.ts"]
+  "include": ["src", "server", "modern.config.ts"]
 }
