--- tests/integration/server-hook/request/.eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  extends: ['@modern-js'],
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: ['./tsconfig.json'],
-  },
-  ignorePatterns: ['./modern.config.ts'],
-};
