+++ tests/integration/server-hook/request/tsconfig.json
@@ -2,12 +2,12 @@
   "extends": "@modern-js/tsconfig/base",
   "compilerOptions": {
     "declaration": false,
-    "types": ["jest", "node"],
     "jsx": "preserve",
     "baseUrl": "./",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@shared/*": ["./shared/*"]
     }
   },
-  "include": ["src", "test", "server", "modern.config.ts"]
+  "include": ["src", "server", "modern.config.ts"]
 }
