+++ tests/integration/server-hook/response/tsconfig.json
@@ -9,5 +9,5 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src", "test"]
+  "include": ["src"]
 }
