+++ tests/integration/server-hook/router/tsconfig.json
@@ -9,5 +9,5 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src", "test"]
+  "include": ["src"]
 }
