+++ tests/integration/server-hook/template/tsconfig.json
@@ -9,5 +9,5 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src", "test"]
+  "include": ["src"]
 }
